Atualiza quantidades das tabelas de INT para DECIMAL

+++ sql/DiarioMotorista_Receitas.sql
@@ -3,7 +3,7 @@ CREATE DATABASE DiarioMotorista;
 USE DiarioMotorista;
 CREATE TABLE receitas (
         receita_id INT AUTO_INCREMENT NOT NULL,
-        receita_quantidade INT (3) NOT NULL,
+        receita_quantidade DECIMAL (10,2) NOT NULL,
         receita_origem VARCHAR (10) NOT NULL,
         receita_data DATE NOT NULL,
         PRIMARY KEY(receita_id)
